moves airline log for consistency
